📖 Message parsing, Orchestration logic skeleton (#7)
This change updates the `IConnection` interface to match the protocol changes in #6. These updates meant that `FtlConnection` and `MockConnection` needed updates as well. The `FtlConnectionUnitTests` have been updated to match.

Also introduced here is the beginning skeleton of the `Orchestrator` logic which will take in `IConnection` messages, determine how to route streams, and send out messages to other `IConnection` nodes to orchestrate those routes. The logic is not yet complete, but a skeleton is there to provide the framework needed to build that logic, and a good test fixture has been created to facilitate testing of `Orchestrator` logic.