🙋‍♂️ Introduce header-only client library (#13)
This change does a bit of refactoring and introduces `FtlOrchestrationClient` for use by other consumers to connect to the FTL Orchestration Service. All the headers required for clients to include have been moved to `inc/` (and in some cases changed to be header-only). 

A simple functional test to validate that clients can connect to the service has been added in `test/functional/FunctionalTests.cpp`. In the future, more functional tests should be added here to test common scenarios between orchestration clients and the orchestration service.